src: gpu: intel: binary: widen offset calc with off_t

+++ src/gpu/intel/binary/simple.cl
@@ -24,7 +24,7 @@ __kernel void simple_binary(__global DATA_T *src0, __global DATA_T *src1,
 #endif
         __global DST_DATA_T *dst POST_OP_ARGS, __global float *src0_scale,
         __global float *src1_scale) {
-    int off = GWS_GET_IDX();
+    off_t off = GWS_GET_IDX();
 
     float tmp_src0 = SRC0_TO_FLOAT(src0[off]);
     float tmp_src1 = SRC1_TO_FLOAT(src1[off]);
@@ -76,30 +76,30 @@ __kernel void simple_binary(__global SRC0_DATA_T *src0,
             = {dims0[0], dims0[1], dims0[2], dims0[3], dims0[4], dims0[5]};
     int d1_init = GWS_GET_D1();
 
-    int dst_off = DST_OFF(
+    off_t dst_off = DST_OFF(
             dims0[0], dims0[1], dims0[2], dims0[3], dims0[4], dims0[5]);
 #if TENSOR_OP
-    int src0_off = SRC0_OFF(
+    off_t src0_off = SRC0_OFF(
             dims0[0], dims0[1], dims0[2], dims0[3], dims0[4], dims0[5]);
-    int src1_off = SRC1_OFF(
+    off_t src1_off = SRC1_OFF(
             dims0[0], dims0[1], dims0[2], dims0[3], dims0[4], dims0[5]);
 
 #if IS_TERNARY
-    int src2_off = SRC2_OFF(
+    off_t src2_off = SRC2_OFF(
             dims0[0], dims0[1], dims0[2], dims0[3], dims0[4], dims0[5]);
 #endif
 
 #else
-    int src0_off
+    off_t src0_off
             = SRC0_OFF(dims0[0] * !SRC0_BCAST_DIM0, dims0[1] * !SRC0_BCAST_DIM1,
                     dims0[2] * !SRC0_BCAST_DIM2, dims0[3] * !SRC0_BCAST_DIM3,
                     dims0[4] * !SRC0_BCAST_DIM4, dims0[5] * !SRC0_BCAST_DIM5);
-    int src1_off
+    off_t src1_off
             = SRC1_OFF(dims0[0] * !SRC1_BCAST_DIM0, dims0[1] * !SRC1_BCAST_DIM1,
                     dims0[2] * !SRC1_BCAST_DIM2, dims0[3] * !SRC1_BCAST_DIM3,
                     dims0[4] * !SRC1_BCAST_DIM4, dims0[5] * !SRC1_BCAST_DIM5);
 #if IS_TERNARY
-    int src2_off
+    off_t src2_off
             = SRC2_OFF(dims0[0] * !SRC2_BCAST_DIM0, dims0[1] * !SRC2_BCAST_DIM1,
                     dims0[2] * !SRC2_BCAST_DIM2, dims0[3] * !SRC2_BCAST_DIM3,
                     dims0[4] * !SRC2_BCAST_DIM4, dims0[5] * !SRC2_BCAST_DIM5);
